test(rewrite_functional) clean up bootstrap

+++ tests/rewrite_functional/_bootstrap.php
@@ -2,21 +2,7 @@
 
 use Codeception\Event\SuiteEvent;
 use Codeception\Event\TestEvent;
-use Codeception\Module\WPFilesystem;
 use function tad\WPBrowser\addListener;
-
-// This function will get around re-configuration or late configuration settings.
-function tec_canonical_url_mu_plugin_path( WPFilesystem $fs ): string {
-	static $pathname;
-
-	if ( $pathname ) {
-		return $pathname;
-	}
-
-	$pathname = $fs->_getConfig( 'mu-plugins' ) . 'tec-canonical-url-service.php';
-
-	return $pathname;
-}
 
 /*
  * Ensure the tec-canonical/url plugin is placed.
@@ -32,12 +18,13 @@ addListener( Codeception\Events::TEST_BEFORE, static function ( TestEvent $e ) {
 
 	$placed = true;
 
-	$container = $e->getTest()->getMetadata()->getService( 'di' );
-	/** @var WPFilesystem $fs */
-	$fs = $container->get( WPFilesystem::class );
-
 	$plugin_source_pathname = codecept_data_dir( 'plugins/tec-canonical-url-service.php' );
-	$plugin_dest_pathname   = tec_canonical_url_mu_plugin_path( $fs );
+	$plugin_dest_dir        = dirname( __DIR__, 4 ) . '/mu-plugins';
+	$plugin_dest_pathname   = $plugin_dest_dir . '/tec-canonical-url-service.php';
+
+	if ( ! is_dir( $plugin_dest_dir ) && mkdir( $plugin_dest_dir, 0777, true ) && is_dir( $plugin_dest_dir ) ) {
+		throw new \RuntimeException( "Could not create directory $plugin_dest_dir" );
+	}
 
 	if ( file_exists( $plugin_dest_pathname ) && ! unlink( $plugin_dest_pathname ) ) {
 		throw new RuntimeException( "Could not remove file $plugin_dest_pathname." );
@@ -68,9 +55,10 @@ addListener( Codeception\Events::TEST_BEFORE, static function ( TestEvent $e ) {
 
 // When the suite is done, remove the mu-plugin file.
 addListener( Codeception\Events::SUITE_AFTER, static function ( SuiteEvent $e ) {
-	$fs       = $e->getSuite()->getModules()['WPFilesystem'];
-	$pathname = tec_canonical_url_mu_plugin_path( $fs );
-	if ( file_exists( $pathname ) ) {
-		$fs->deleteFile( $pathname );
+	$fs                   = $e->getSuite()->getModules()['WPFilesystem'];
+	$plugin_dest_dir      = dirname( __DIR__, 4 ) . '/mu-plugins';
+	$plugin_dest_pathname = $plugin_dest_dir . '/tec-canonical-url-service.php';
+	if ( file_exists( $plugin_dest_pathname ) ) {
+		$fs->deleteFile( $plugin_dest_pathname );
 	}
 } );
